feat: opt-out of initial service call for fpm project

+++ generators/uimodule/ui5Libs.js
@@ -77,19 +77,23 @@ export default class extends Generator {
 				name: "sap.fe.templates"
 			})
 			ui5Yaml.server.customMiddleware = ui5Yaml.server.customMiddleware.filter(middleware => middleware.name !== "sap-fe-mockserver")
-			const ui5YamlMock = yaml.parse(fs.readFileSync(this.destinationPath("ui5-mock.yaml")).toString())
-			ui5YamlMock.framework = {
-				name: "SAPUI5",
-				version: dependencies["SAPUI5"],
-				libraries: [
-					{ name: "sap.m" },
-					{ name: "sap.ui.core" },
-					{ name: "themelib_sap_horizon" },
-					{ name: "sap.fe.templates" }
-				]
+			const ui5YamlMockPath = this.destinationPath("ui5-mock.yaml")
+			let ui5YamlMock
+			if (fs.existsSync(ui5YamlMockPath)) {
+				ui5YamlMock = yaml.parse(fs.readFileSync(this.destinationPath("ui5-mock.yaml")).toString())
+				ui5YamlMock.framework = {
+					name: "SAPUI5",
+					version: dependencies["SAPUI5"],
+					libraries: [
+						{ name: "sap.m" },
+						{ name: "sap.ui.core" },
+						{ name: "themelib_sap_horizon" },
+						{ name: "sap.fe.templates" }
+					]
+				}
+				deleteProxyToUI5(ui5YamlMock)
+				fs.writeFileSync(this.destinationPath("ui5-mock.yaml"), yaml.stringify(ui5YamlMock))
 			}
-			deleteProxyToUI5(ui5YamlMock)
-			fs.writeFileSync(this.destinationPath("ui5-mock.yaml"), yaml.stringify(ui5YamlMock))
 		}
 
 		fs.writeFileSync(this.destinationPath("ui5.yaml"), yaml.stringify(ui5Yaml))
